Create App configuration (#21)
## Summary

This introduces the ability to load app configuration from
`HOMEDIR/.lazynx/lazynx.json` file. At the moment can be used to define
logs file directory.

## Related Issue

Related to #15 

## Type

<!-- What type of change is this? -->

- [ ] Bug fix
- [x] Feature
- [ ] Documentation
- [ ] Refactoring
- [ ] Other: <!-- specify -->

## Testing

I added a global configuration with `.lazynx.json` specifying some logs
path and was able to load the config and write the logs properly for
both the TUI and the NXClient

## Additional Notes

**Workspace Configuration**

Ideally should support load configuration from the workspace path. In
order to do this, there are some considerations related to the logs when
prompt workspace path, those will need to be removed or change the way
are written since we need the workspace path before creating the
configuration type so all the logs in the process can't be written.

PS: I did open a previous PR with the same changes but wrong commit
Author, apologies for that.